feat: implement JWT authentication and permissions using Auth0

+++ test_app.py
@@ -24,12 +24,13 @@ class FundMyFunTestCase(unittest.TestCase):
     def setUp(self):
         """Executed before reach test"""
         load_dotenv()
-        self.database_name = os.getenv('DBNAME')
-        self.database_user = os.getenv('DBUSER')
-        self.database_pw = os.getenv('DBPW')
-        self.database_host = os.getenv('DBHOST')
-        self.database_path = "postgresql+psycopg2://{}:{}@{}/{}".format(self.database_user, self.database_pw,
-                                                                        self.database_host, self.database_name)
+        # self.database_name = os.getenv('DBNAME')
+        # self.database_user = os.getenv('DBUSER')
+        # self.database_pw = os.getenv('DBPW')
+        # self.database_host = os.getenv('DBHOST')
+        # self.database_path = "postgresql+psycopg2://{}:{}@{}/{}".format(self.database_user, self.database_pw,
+        #                                                                 self.database_host, self.database_name)
+        self.database_path = os.getenv('DATABASE_URL')
         self.app = create_app(self.database_path)
         self.client = self.app.test_client
 
@@ -38,6 +39,9 @@ class FundMyFunTestCase(unittest.TestCase):
         with self.app.app_context():
             self.db = SQLAlchemy()
             self.db.init_app(self.app)
+
+        self.JWT_Fundraiser = os.getenv('JWT_Fundraiser')
+        self.JWT_Contributor = os.getenv('JWT_Contributor')
 
     def tearDown(self):
         """Executed after reach test"""
@@ -53,7 +57,8 @@ class FundMyFunTestCase(unittest.TestCase):
 
     def test_post_money_pot(self):
         res = self.client().post('/money_pots', json={"title": "test create money_pot", "description": "This is a test.",
-                                                      "target": "100", "owner_id": "1"})
+                                                      "target": "100", "owner_id": "1"},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -63,11 +68,20 @@ class FundMyFunTestCase(unittest.TestCase):
 
     def test_400_post_money_pot(self):
         res = self.client().post('/money_pots', json={"title": None, "description": "This is a failing test.",
-                                                      "target": "100", "owner_id": None})
+                                                      "target": "100", "owner_id": None},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 400)
         self.assertEqual(data['message'], 'Bad request.')
+
+    def test_401_post_money_pot(self):
+        res = self.client().post('/money_pots', json={"title": "test create money_pot", "description": "This is a test.",
+                                                      "target": "100", "owner_id": "1"},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
 
     def test_get_money_pots(self):
         res = self.client().get('/money_pots')
@@ -78,78 +92,117 @@ class FundMyFunTestCase(unittest.TestCase):
 
     def test_get_money_pot_details(self):
         id = self.test_post_money_pot()
-        res = self.client().get('/money_pots/' + str(id))
+        res = self.client().get('/money_pots/' + str(id), headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['money_pot_id'], id)
 
     def test_404_get_money_pot_details(self):
-        res = self.client().get('/money_pots/1000')
+        res = self.client().get('/money_pots/1000', headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['message'], 'Resource not found.')
 
+    def test_401_get_money_pot_details(self):
+        id = self.test_post_money_pot()
+        res = self.client().get('/money_pots/' + str(id), headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
+
     def test_delete_money_pot(self):
         id = self.test_post_money_pot()
-        res = self.client().delete('/money_pots/'+str(id))
+        res = self.client().delete('/money_pots/' + str(id), headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['money_pot_id'], id)
 
     def test_404_delete_money_pot(self):
-        res = self.client().delete('/money_pots/1000')
+        res = self.client().delete('/money_pots/1000', headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['message'], 'Resource not found.')
 
+    def test_401_delete_money_pot(self):
+        id = self.test_post_money_pot()
+        res = self.client().delete('/money_pots/' + str(id), headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
+
     def test_patch_money_pot(self):
         id = self.test_post_money_pot()
-        res = self.client().patch('/money_pots/' + str(id), json={"description": "This is the new description."})
+        res = self.client().patch('/money_pots/' + str(id), json={"description": "This is the new description."},
+                                  headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['money_pot_id'], id)
 
     def test_404_patch_money_pot(self):
-        res = self.client().patch('/money_pots/1000', json={"description": "This is the new description."})
+        res = self.client().patch('/money_pots/1000', json={"description": "This is the new description."},
+                                  headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['message'], 'Resource not found.')
 
+    def test_401_patch_money_pot(self):
+        id = self.test_post_money_pot()
+        res = self.client().patch('/money_pots/' + str(id), json={"description": "This is the new description."},
+                                  headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
+
     # TEST PLEDGES ENDPOINTS
     def test_post_pledge(self):
-        res = self.client().post('/pledges', json={"user_id": "2", "money_pot_id": "1", "amount": "10"})
+        res = self.client().post('/pledges', json={"user_id": "2", "money_pot_id": "1", "amount": "10"},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['pledge_id'])
 
     def test_400_post_pledge(self):
-        res = self.client().post('/pledges', json={"user_id": None, "money_pot_id": None, "amount": "10"})
+        res = self.client().post('/pledges', json={"user_id": None, "money_pot_id": None, "amount": "10"},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 400)
         self.assertEqual(data['message'], 'Bad request.')
 
+    def test_401_post_pledge(self):
+        res = self.client().post('/pledges', json={"user_id": "2", "money_pot_id": "1", "amount": "10"},
+                                 headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
+
     # TEST USERS ENDPOINTS
     def test_get_user_details(self):
-        res = self.client().get('/users/2')
+        res = self.client().get('/users/2', headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertTrue(data['user_details'])
 
     def test_404_get_user_details(self):
-        res = self.client().get('/users/1000')
+        res = self.client().get('/users/1000', headers={'Authorization': 'Bearer ' + self.JWT_Contributor})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['message'], 'Resource not found.')
+
+    def test_401_get_user_details(self):
+        res = self.client().get('/users/2', headers={'Authorization': 'Bearer ' + self.JWT_Fundraiser})
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
 
 if __name__ == "__main__":
     unittest.main()
